pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "officereader-mcp"
version = "2.0.0"
description = "MCP server for converting Office documents (Word, Excel, PowerPoint) to Markdown with image extraction and optimization"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Asunainlove" },
    { email = "239791209+Asunainlove@users.noreply.github.com" }
]
maintainers = [
    { name = "Asunainlove" },
    { email = "239791209+Asunainlove@users.noreply.github.com" }
]
keywords = ["mcp", "docx", "doc", "xlsx", "xls", "pptx", "ppt", "markdown", "converter", "office", "claude", "excel", "powerpoint", "word", "office-documents"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "python-docx>=1.1.0",
    "mammoth>=1.6.0",
    "Pillow>=10.0.0",
    "markdownify>=0.11.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
officereader-mcp = "officereader_mcp.server:main"

[project.urls]
Homepage = "https://github.com/Asunainlove/office-reader-mcp"
Repository = "https://github.com/Asunainlove/office-reader-mcp"
Issues = "https://github.com/Asunainlove/office-reader-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/officereader_mcp"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
